+++ test/utils/CatalystClientBuilder.spec.ts
@@ -1,105 +1,113 @@
-import { getUpdatedApprovedListWithoutQueryingContract } from '../../src/utils/CatalystClientBuilder'
+import { clientConnectedToCatalystIn } from '../../src/utils/CatalystClientBuilder'
+import * as catalystList from '../../src/utils/catalystList'
+// import { CatalystClient } from '../../src/CatalystClient'
+import { HealthStatus } from 'dcl-catalyst-commons'
+import * as common from '../../src/utils/common'
+import { CatalystClient } from 'CatalystClient'
 
-describe('Catalyst Client Builder', () => {
-  describe('getUpdatedApprovedListWithoutQueryingContract', () => {
-    const ORIGIN = 'origin'
-    const REQUIRED_LISTS = 2
-    const [ADDRESS1, ADDRESS2, ADDRESS3] = ['http://test1.com', 'http://test2.com', 'http://test3.com']
+const server1 = 'server1'
+const healthyServer = 'server2'
+const server3 = 'server3'
 
-    it('When the amount of known servers on the list is less than required, then no fetching is performed and nothing is returned', async () => {
-      let fetchCount = 0
+const mockedServerList = [server1, healthyServer, server3]
+jest.mock('../../src/CatalystClient', () => ({
+  CatalystClient: jest
+    .fn()
+    .mockImplementationOnce(() => {
+      return {
+        fetchPeerHealth: jest.fn().mockReturnValueOnce({
+          comms: HealthStatus.UNHEALTHY,
+          lambdas: HealthStatus.HEALTHY,
+          content: HealthStatus.DOWN
+        })
+      }
+    })
+    .mockImplementationOnce(() => {
+      return {
+        fetchPeerHealth: jest.fn().mockReturnValueOnce({
+          comms: HealthStatus.HEALTHY,
+          lambdas: HealthStatus.HEALTHY,
+          content: HealthStatus.HEALTHY
+        })
+      }
+    })
+    .mockImplementationOnce(() => {
+      return {
+        fetchPeerHealth: jest.fn().mockReturnValueOnce({
+          comms: HealthStatus.UNHEALTHY,
+          lambdas: HealthStatus.HEALTHY,
+          content: HealthStatus.DOWN
+        })
+      }
+    })
+    .mockImplementationOnce(() => {
+      return {
+        fetchPeerHealth: jest.fn().mockReturnValueOnce({
+          comms: HealthStatus.HEALTHY,
+          lambdas: HealthStatus.HEALTHY,
+          content: HealthStatus.HEALTHY
+        })
+      }
+    })
+}))
 
-      const result = await calculateList({
-        list: [{ address: ADDRESS1 }],
-        fetchApprovedCatalysts: () => {
-          fetchCount++
-          return Promise.resolve([])
-        }
-      })
+describe('clientConnectedToCatalystIn', () => {
+  let getUpdatedApprovedListWithoutQueryingContractSpy: jest.SpyInstance
+  let getApprovedListFromContractSpy: jest.SpyInstance
+  let shuffleArraySpy: jest.SpyInstance
 
-      expect(fetchCount).toEqual(0)
-      expect(result).toBeUndefined()
+  describe('when the noContractList is empty', () => {
+    beforeEach(() => {
+      getUpdatedApprovedListWithoutQueryingContractSpy = jest.spyOn(
+        catalystList,
+        'getUpdatedApprovedListWithoutQueryingContract'
+      )
+      getApprovedListFromContractSpy = jest.spyOn(catalystList, 'getApprovedListFromContract')
+      shuffleArraySpy = jest.spyOn(common, 'shuffleArray').mockReturnValueOnce(mockedServerList)
+      getUpdatedApprovedListWithoutQueryingContractSpy.mockReturnValueOnce(Promise.resolve(undefined))
+      getApprovedListFromContractSpy.mockReturnValue(Promise.resolve(mockedServerList))
     })
 
-    it(`When one of the first N servers doesn't respond, then next one is queried`, async () => {
-      let fetchCount = 0
-
-      await calculateList({
-        list: [{ address: ADDRESS1 }, { address: ADDRESS2 }, { address: ADDRESS3 }],
-        fetchApprovedCatalysts: () => {
-          if (fetchCount++ === 0) {
-            return Promise.resolve(undefined)
-          }
-          return Promise.resolve([ADDRESS1])
-        }
-      })
-
-      expect(fetchCount).toEqual(3)
+    afterEach(() => {
+      getUpdatedApprovedListWithoutQueryingContractSpy.mockRestore()
+      getApprovedListFromContractSpy.mockRestore()
+      shuffleArraySpy.mockClear()
+      shuffleArraySpy.mockRestore()
     })
 
-    it(`When not enough servers responded with a valid list, then nothing is returned`, async () => {
-      let fetchCount = 0
+    it('should use get the approved list from the DAO', async () => {
+      await clientConnectedToCatalystIn('mainnet', '')
+      expect(shuffleArraySpy).toHaveBeenCalledTimes(1)
+      expect(shuffleArraySpy).toHaveBeenCalledWith(mockedServerList)
+    })
+  })
 
-      const result = await calculateList({
-        list: [{ address: ADDRESS1 }, { address: ADDRESS2 }, { address: ADDRESS3 }],
-        fetchApprovedCatalysts: () => {
-          fetchCount++
-          return Promise.resolve(undefined)
-        }
-      })
+  describe('when the noContractList is not empty', () => {
+    let result: CatalystClient
 
-      expect(fetchCount).toEqual(3)
-      expect(result).toBeUndefined()
+    beforeEach(async () => {
+      getUpdatedApprovedListWithoutQueryingContractSpy = jest.spyOn(
+        catalystList,
+        'getUpdatedApprovedListWithoutQueryingContract'
+      )
+      getApprovedListFromContractSpy = jest.spyOn(catalystList, 'getApprovedListFromContract')
+      shuffleArraySpy = jest.spyOn(common, 'shuffleArray').mockReturnValueOnce(mockedServerList)
+      getUpdatedApprovedListWithoutQueryingContractSpy.mockReturnValue(Promise.resolve(mockedServerList))
+      getApprovedListFromContractSpy.mockReturnValue(Promise.resolve(undefined))
+
+      result = await clientConnectedToCatalystIn('mainnet', '')
     })
 
-    it(`When the intersection is empty, then nothing is returned`, async () => {
-      const result = await calculateList({
-        list: [{ address: ADDRESS1 }, { address: ADDRESS2 }, { address: ADDRESS3 }],
-        fetchApprovedCatalysts: (catalystUrl) => {
-          return Promise.resolve([catalystUrl])
-        }
-      })
-
-      expect(result).toBeUndefined()
+    afterEach(() => {
+      getUpdatedApprovedListWithoutQueryingContractSpy.mockRestore()
+      getApprovedListFromContractSpy.mockRestore()
+      shuffleArraySpy.mockRestore()
     })
 
-    it(`When there are enough updated lists, then the intersection is returned`, async () => {
-      let fetchCount = 0
-
-      const result = await calculateList({
-        list: [{ address: ADDRESS1 }, { address: ADDRESS2 }],
-        fetchApprovedCatalysts: async () => {
-          switch (fetchCount++) {
-            case 0:
-              return [ADDRESS1, ADDRESS2]
-            case 1:
-              return [ADDRESS2, ADDRESS3]
-          }
-          return undefined
-        }
-      })
-
-      expect(fetchCount).toEqual(2)
-      expect(result).toEqual([ADDRESS2])
+    it('should use noContractList as the list of servers and return a result', () => {
+      expect(shuffleArraySpy).toHaveBeenCalledTimes(1)
+      expect(shuffleArraySpy).toHaveBeenCalledWith(mockedServerList)
+      expect(result).toBeDefined()
     })
-
-    // describe('isServerUp', () => {
-    //   const
-    // })
-
-    function calculateList({
-      list,
-      fetchApprovedCatalysts
-    }: {
-      list: { address: string }[]
-      fetchApprovedCatalysts: (catalystUrl: string) => Promise<string[] | undefined>
-    }) {
-      return getUpdatedApprovedListWithoutQueryingContract({
-        origin: ORIGIN,
-        requiredLists: REQUIRED_LISTS,
-        preKnownServers: { list },
-        fetchApprovedCatalysts
-      })
-    }
   })
 })
